Completed ssh console resizing support on Windows
