+++ ssh/io.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"io"
-	"runtime"
 	"sync"
 
 	"golang.org/x/crypto/ssh"
@@ -47,34 +46,8 @@ func getFd(reader io.Reader) (fd int, ok bool) {
 	return fd, term.IsTerminal(fd)
 }
 
-func (s *SessionIO) getAndWatchSize(ctx context.Context, sess *ssh.Session) (int, int, error) {
-	var (
-		fd int
-		ok bool
-	)
-	if runtime.GOOS == "windows" {
-		// TODO(Ali): Hardcoded stdout instead of pulling it from the SessionIO because it's
-		//            wrapped in multiple wrapper types.
-		fd = 1
-		ok = true
-	} else {
-		fd, ok = getFd(s.Stdin)
-	}
-	if !ok {
-		return 0, 0, errors.New("could not get console handle")
-	}
-
-	width, height, err := term.GetSize(fd)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	go watchWindowSize(ctx, fd, sess)
-
-	return width, height, nil
-}
-
 func (s *SessionIO) attach(ctx context.Context, sess *ssh.Session, cmd string) error {
+
 	if s.AllocPTY {
 		width, height := DefaultWidth, DefaultHeight
 
@@ -85,6 +58,7 @@ func (s *SessionIO) attach(ctx context.Context, sess *ssh.Session, cmd string) e
 			}
 			defer term.Restore(fd, state)
 		}
+
 
 		if w, h, err := s.getAndWatchSize(ctx, sess); err == nil {
 			width, height = w, h
